refactor: samefy the date utils namespace to the other utils
