+++ src/TMTProductizer/Utils/DateUtils.cs
@@ -1,12 +1,12 @@
 
-namespace TMTProductizer.Utils.DateUtils;
+namespace TMTProductizer.Utils;
 
 public static class DateUtils
 {
     public static DateTime UnixTimeStampToDateTime(double unixTimeStamp)
     {
         // Unix timestamp is seconds past epoch
-        DateTime dtDateTime = new DateTime(1970,1,1,0,0,0,0, System.DateTimeKind.Utc);
+        DateTime dtDateTime = new DateTime(1970, 1, 1, 0, 0, 0, 0, System.DateTimeKind.Utc);
         dtDateTime = dtDateTime.AddSeconds(unixTimeStamp).ToLocalTime();
         return dtDateTime;
     }
